Reorganise and remove more stuff from common.
