+++ ocean_lib/assets/asset.py
@@ -12,12 +12,12 @@ from enforce_typing import enforce_types
 from eth_account.account import Account
 from eth_utils import add_0x_prefix
 from ocean_lib.assets.credentials import AddressCredential
+from ocean_lib.assets.did import did_to_id
 from ocean_lib.common.agreements.consumable import ConsumableCodes
 from ocean_lib.common.agreements.service_types import ServiceTypes
-from ocean_lib.common.did import did_to_id
-from ocean_lib.common.utils.utilities import get_timestamp
 from ocean_lib.data_provider.data_service_provider import DataServiceProvider
 from ocean_lib.services.service import Service
+from ocean_lib.utils.utilities import get_timestamp
 from ocean_lib.web3_internal.wallet import Wallet
 
 logger = logging.getLogger("ddo")
