--- ocean_lib/common/utils/__init__.py
@@ -1,4 +1,0 @@
-#
-# Copyright 2021 Ocean Protocol Foundation
-# SPDX-License-Identifier: Apache-2.0
-#
