+++ ocean_lib/utils/test/test_utilities.py
@@ -2,7 +2,7 @@
 # Copyright 2021 Ocean Protocol Foundation
 # SPDX-License-Identifier: Apache-2.0
 #
-from ocean_lib.common.utils import utilities
+from ocean_lib.utils import utilities
 
 
 def test_convert():
